requirements.txt
rapidfuzz
pandas
numpy
pyarrow
sqlalchemy
python-dotenv
mariadb
urllib3
certifi